requirements.txt
absl-py==1.0.0
asgiref==3.5.0
astunparse==1.6.3
autopep8==1.6.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.10
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.11.1
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
mysqlclient==2.1.0

oauthlib==3.2.0
opt-einsum==3.3.0
Pillow==9.1.0
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
resumable-urlretrieve==0.1.6
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-cpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
toml==0.10.2
typing_extensions==4.1.1
tzdata==2021.5
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.7.0
